Preparation for Signing V4 implementation
- Put nested classes into separate files
- Create a private ISigner interface to simplify the implementation
- Introduce two implementations: V2Signer and V4Signer, the latter just throwing
- Implement public methods by choosing an appropriate ISigner

All tests still pass, but this isn't testing anything about V4 at all yet, as it's unimplemented.
