+++ apis/Google.Cloud.Storage.V1/Google.Cloud.Storage.V1/UrlSigner.V4Signer.cs
@@ -1,0 +1,53 @@
+﻿// Copyright 2018 Google LLC
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     https://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Collections.Generic;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Google.Cloud.Storage.V1
+{
+    public sealed partial class UrlSigner
+    {
+        private sealed class V4Signer : ISigner
+        {
+            public string Sign(
+                string bucket,
+                string objectName,
+                DateTimeOffset? expiration,
+                HttpMethod requestMethod,
+                Dictionary<string, IEnumerable<string>> requestHeaders,
+                Dictionary<string, IEnumerable<string>> contentHeaders,
+                IBlobSigner blobSigner)
+            {
+                throw new NotImplementedException();
+            }
+
+            public Task<string> SignAsync(
+                string bucket,
+                string objectName,
+                DateTimeOffset? expiration,
+                HttpMethod requestMethod,
+                Dictionary<string, IEnumerable<string>> requestHeaders,
+                Dictionary<string, IEnumerable<string>> contentHeaders,
+                CancellationToken cancellationToken,
+                IBlobSigner blobSigner)
+            {
+                throw new NotImplementedException();
+            }
+        }
+    }
+}
